requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-mail==0.9.1
PyMySQL==1.1.0
redis==5.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
opencv-python==4.9.0.80
mediapipe==0.10.14
deepface==0.0.93
tensorflow==2.15.0
tf-keras==2.15.0
numpy==1.26.4
scipy>=1.11.0
scikit-image>=0.21.0
pillow>=10.0.0
PyJWT==2.8.0
cryptography==41.0.0
protobuf==4.25.3